move more symbolic to ops.py [pr]
